rag-python/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
numpy
psycopg2-binary
python-dotenv